# api/imagebuilder/v1alpha1/openapi.yaml
openapi: 3.0.1
info:
  title: Flight Control Image Builder API
  version: v1alpha1
  description: |
    Image Builder service for Flight Control.
    This service manages image build resources for building and pushing container images.
  contact:
    name: The Flight Control Team
    url: https://flightctl.io
    email: team@flightctl.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
  /api/v1/imagebuilds:
    x-resource: imagebuilds
    get:
      tags:
        - imagebuild
      description: List ImageBuild resources.
      operationId: listImageBuilds
      parameters:
        - name: labelSelector
          in: query
          description: A selector to restrict the list of returned objects by their labels.
          schema:
            type: string
        - name: fieldSelector
          in: query
          description: A selector to restrict the list of returned objects by their fields.
          schema:
            type: string
        - name: limit
          in: query
          description: The maximum number of results returned in the list response.
          schema:
            type: integer
            format: int32
        - name: continue
          in: query
          description: An optional parameter to query more results from the server.
          schema:
            type: string
        - name: withExports
          in: query
          description: If true, includes related ImageExport resources in each ImageBuild in the response.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuildList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    post:
      tags:
        - imagebuild
      description: Create an ImageBuild resource.
      operationId: createImageBuild
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageBuild'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuild'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/imagebuilds/{name}:
    x-resource: imagebuilds
    get:
      tags:
        - imagebuild
      description: Get an ImageBuild resource.
      operationId: getImageBuild
      parameters:
        - name: name
          in: path
          description: The name of the ImageBuild resource.
          required: true
          schema:
            type: string
        - name: withExports
          in: query
          description: If true, includes related ImageExport resources in the response.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuild'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      tags:
        - imagebuild
      description: Delete an ImageBuild resource.
      operationId: deleteImageBuild
      parameters:
        - name: name
          in: path
          description: The name of the ImageBuild resource.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuild'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  securitySchemes:
    bearerAuth:
      $ref: '../../core/v1beta1/openapi.yaml#/components/securitySchemes/bearerAuth'
    orgId:
      $ref: '../../core/v1beta1/openapi.yaml#/components/securitySchemes/orgId'
  schemas:
...
    ImageBuild:
      type: object
      description: ImageBuild represents a build request for a container image.
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '../../core/v1beta1/openapi.yaml#/components/schemas/ObjectMeta'
        spec:
          $ref: '#/components/schemas/ImageBuildSpec'
        status:
          $ref: '#/components/schemas/ImageBuildStatus'
        imageexports:
          type: array
          description: Array of ImageExport resources that reference this ImageBuild. Only populated when withExports query parameter is true.
          items:
            $ref: '#/components/schemas/ImageExport'
      required:
        - apiVersion
        - kind
        - metadata
        - spec
    ImageBuildList:
      type: object
      description: ImageBuildList is a list of ImageBuild resources.
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '../../core/v1beta1/openapi.yaml#/components/schemas/ListMeta'
        items:
          type: array
          description: 'List of ImageBuild resources.'
          items:
            $ref: '#/components/schemas/ImageBuild'
      required:
        - apiVersion
        - kind
        - metadata
        - items
    ImageBuildSpec:
      type: object
      description: ImageBuildSpec describes the specification for an image build.
      properties:
        source:
          $ref: '#/components/schemas/ImageBuildSource'
        destination:
          $ref: '#/components/schemas/ImageBuildDestination'
        binding:
          $ref: '#/components/schemas/ImageBuildBinding'
        userConfiguration:
          $ref: '#/components/schemas/ImageBuildUserConfiguration'
      required:
        - source
        - destination
        - binding
    ImageBuildSource:
      type: object
      description: ImageBuildSource specifies the source image for the build.
      properties:
        repository:
          type: string
          description: The name of the Repository resource of type OCI containing the source image.
        imageName:
          type: string
          description: The name of the source image.
        imageTag:
          type: string
          description: The tag of the source image.
      required:
        - repository
        - imageName
        - imageTag
    ImageBuildDestination:
      type: object
      description: ImageBuildDestination specifies the destination for the built image.
      properties:
        repository:
          type: string
          description: The name of the Repository resource of type OCI to push the built image to.
        imageName:
          type: string
          description: The name of the output image.
        imageTag:
          type: string
          description: The tag of the output image.
      required:
        - repository
        - imageName
        - imageTag
    ImageBuildBinding:
      description: ImageBuildBinding specifies binding configuration for the build.
      oneOf:
        - $ref: '#/components/schemas/EarlyBinding'
        - $ref: '#/components/schemas/LateBinding'
      discriminator:
        propertyName: type
        mapping:
          early: '#/components/schemas/EarlyBinding'
          late: '#/components/schemas/LateBinding'
    ImageBuildUserConfiguration:
      type: object
      description: ImageBuildUserConfiguration specifies user configuration for the build.
      properties:
        username:
          type: string
          description: The username for the user configuration.
        publickey:
          type: string
          description: The public key for the user configuration.
      required:
        - username
        - publickey
    EarlyBinding:
      type: object
      description: Early binding configuration - embeds certificate in the image.
      required:
        - type
      properties:
        type:
          type: string
          enum: [early]
          description: The type of binding.
    LateBinding:
      type: object
      description: Late binding configuration - device binds at first boot.
      required:
        - type
      properties:
        type:
          type: string
          enum: [late]
          description: The type of binding.
...
    ImageBuildStatus:
      type: object
      description: ImageBuildStatus represents the current status of an ImageBuild.
      properties:
        conditions:
          type: array
          description: Current conditions of the ImageBuild.
          items:
            $ref: '#/components/schemas/ImageBuildCondition'
        imageReference:
          type: string
          description: The full image reference of the built image (e.g., quay.io/org/imagename:tag).
        architecture:
          type: string
          description: The architecture of the built image.
        manifestDigest:
          type: string
          description: The digest of the built image manifest.
        lastSeen:
          type: string
          format: date-time
          description: The last time the build was seen (heartbeat).
    ImageBuildCondition:
      description: Condition for ImageBuild resources.
      allOf:
        - $ref: '../../core/v1beta1/openapi.yaml#/components/schemas/ConditionBase'
        - type: object
          required:
            - type
          properties:
            type:
              $ref: '#/components/schemas/ImageBuildConditionType'
    ImageBuildConditionType:
      type: string
      description: Type of ImageBuild condition.
      enum:
        - Ready
      x-enum-varnames:
        - ImageBuildConditionTypeReady
...
    ImageExport:
      type: object
      description: ImageExport represents an export request to convert and push images to different formats.
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '../../core/v1beta1/openapi.yaml#/components/schemas/ObjectMeta'
        spec:
          $ref: '#/components/schemas/ImageExportSpec'
        status:
          $ref: '#/components/schemas/ImageExportStatus'
      required:
        - apiVersion
        - kind
        - metadata
        - spec
...
    ImageExportSpec:
      type: object
      description: ImageExportSpec describes the specification for an image export.
      properties:
        source:
          $ref: '#/components/schemas/ImageExportSource'
        format:
          $ref: '#/components/schemas/ExportFormatType'
      required:
        - source
        - format
    ImageExportSource:
      description: ImageExportSource specifies the source image for the export.
      oneOf:
        - $ref: '#/components/schemas/ImageBuildRefSource'
      discriminator:
        propertyName: type
        mapping:
          imageBuild: '#/components/schemas/ImageBuildRefSource'
    ImageBuildRefSource:
      type: object
      description: ImageBuildRefSource specifies a source image from an ImageBuild resource.
      required:
        - type
        - imageBuildRef
      properties:
        type:
          type: string
          enum: [imageBuild]
          description: The type of source.
        imageBuildRef:
          type: string
          description: The name of the ImageBuild resource to use as source.
...
    ExportFormatType:
      type: string
      description: The type of format to export the image to.
      enum:
        - vmdk
        - qcow2
        - iso
        - qcow2-disk-container
      x-enum-varnames:
        - ExportFormatTypeVMDK
        - ExportFormatTypeQCOW2
        - ExportFormatTypeISO
        - ExportFormatTypeQCOW2DiskContainer
    ImageExportStatus:
      type: object
      description: ImageExportStatus represents the current status of an ImageExport.
      properties:
        conditions:
          type: array
          description: Current conditions of the ImageExport.
          items:
            $ref: '#/components/schemas/ImageExportCondition'
        manifestDigest:
          type: string
          description: The digest of the exported image manifest for this format.
        lastSeen:
          type: string
          format: date-time
          description: The last time the export was seen (heartbeat).
    ImageExportCondition:
      description: Condition for ImageExport resources.
      allOf:
        - $ref: '../../core/v1beta1/openapi.yaml#/components/schemas/ConditionBase'
        - type: object
          required:
            - type
          properties:
            type:
              $ref: '#/components/schemas/ImageExportConditionType'
    ImageExportConditionType:
      type: string
      description: Type of ImageExport condition.
      enum:
        - Ready
      x-enum-varnames:
        - ImageExportConditionTypeReady
...
    Status:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        code:
          type: integer
          format: int32
          description: Suggested HTTP return code for this status, 0 if not set.
        message:
          type: string
          description: A human-readable description of the status of this operation.
        reason:
          type: string
          description: A machine-readable description of why this operation is in the "Failure" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.
        status:
          type: string
          description: 'Status of the operation. One of: "Success" or "Failure". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.'
      description: Status is a return value for calls that don't return other objects.
      required:
        - apiVersion
        - kind
        - status
        - code
        - reason
        - message